Add 'promise' versions of EventLoopFuture.{and,when}All{succeed,complete}
Motivation:

Most NIO APIs have two versions: one accepting a promise (often
optional) and another which returns a future. Many applications and
frameworks built on top of NIO also use this pattern. To maximise
composability we should try to provide both versions where possible.

Modifications:

Add versions of `andAllSucceed`, `andAllComplete`, `whenAllSucceed`
and `whenAllComplete` which accept an `EventLoopPromise` in place of
an `EventLoop` and defer the implementation of the future versions to
the promise versions.

Result:

- Better composability.
- Fewer allocations if the caller already has a promise.