Nicer shutdown hooks for sockets.
Motivation:

Adjust to a more generic shutdown hook for sockets after PR feedback.

Modification:

Changed the custom cleanup hook to be a more generic shutdownSocket hook
that is implemented in both BaseStreamSocketChannel and PipeChannel.

Result:

A better design for the cleanup hook which may be more generically useful in the future.