Add more synchronous pipeline operations. (#1908)
Motivation:

NIO provides some scope for performing "synchronous" ChannelPipeline
operations. These are only safe to perform from the event loop to which
the pipeline is bound, and they avoid the need to check what thread
context you're running on in release mode.

Several of the built-in channels currently skip the public, thread-safe
API to ChannelPipeline in favour of directly calling into the
synchronous pipeline interface. This is going to cause us some trouble
when we pull ChannelPipeline out into NIOCore.

To that end, we should add the rest of the ChannelPipeline API surface
area to the SyncOperations structure and convert ourselves over to using
them.

Modifications:

- Implemented the pipeline operations on SyncOperations
- Added a test
- Swapped our channels to use the new APIs.

Result:

Better abstraction between ChannelPipeline and the Channels

Co-authored-by: George Barnett <gbarnett@apple.com>