Move UniversalBootstrap to NIOCore.
Motivation:

The UniversalBootstrap abstraction allows you to write code that is
agnostic to the specific event loop and channel implementation. This is
likely to be used in places that don't want to have to express a
concrete requirement for a specific channel or loop, and so properly
belongs in NIOCore.

Modifications:

- Move Universal Bootstrap
- Move TCP convenience options, which are required by the universal
  bootstrap.

Result:

Universal Bootstrap will be in NIOCore.