Add appropriate Sendable conformances.
Motivation:

As part of the adoption of 5.5 there is now some limited amount of
runtime enforcement of thread-safety in the form of the Sendable marker
protocol. A number of NIO's types are safe to share across threads, so
to smooth the path to wider adoption of Sendable we should adopt this
marker protocol where we can.

Modifications:

- Added Sendable to many, many types.

Result:

Users can use NIO in concurrency-aware contexts more easily.